fix oss-fuzz 19896.

+++ qemu/target-arm/translate.c
@@ -11126,7 +11126,7 @@ static void disas_thumb_insn(CPUARMState *env, DisasContext *s) // qq
 
         /* jump to the offset */
         val = (uint32_t)s->pc + 2;
-        offset = ((int32_t)insn << 24) >> 24;
+        offset = ((int32_t)((uint32_t)insn << 24)) >> 24;
         val += offset << 1;
         gen_jmp(s, val);
         break;
@@ -11139,7 +11139,7 @@ static void disas_thumb_insn(CPUARMState *env, DisasContext *s) // qq
         }
         /* unconditional branch */
         val = (uint32_t)s->pc;
-        offset = ((int32_t)insn << 21) >> 21;
+        offset = ((int32_t)((uint32_t)insn << 21)) >> 21;
         val += (offset << 1) + 2;
         gen_jmp(s, val);
         break;
